@@ -4,12 +4,12 @@ import Pojos.CreateUserRequest;
 import Pojos.CreateUserResponse;
 import io.restassured.http.ContentType;
 import org.apache.http.HttpStatus;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
 
-public class ReqResTest01 extends BaseTest{
+public class ReqResTest extends BaseTest{
 
 
 
